drm/fourcc: Add 32/64x4 Tile formats support
Add support for the different tile formats supported by the VCU2.
Each tile format can be interpreted as -
T - Indicates format is tiled
5/6 - tile size of 2^5 = 32 or 2^6 = 64 pixels
M/0/2/4 - Monochrome/420/422/444 Chroma sub sampling
8/A/C - 8/10/12 bit depth

char_per_block will overflow for tiled formats. Hence data type is
changed from u8 to u32.

Tested-by: Neel Gandhi <neel.gandhi@amd.com>
Message-ID: <20260218185731.1769383-5-vishal.sagar@amd.com>
State: pending

+++ include/drm/drm_fourcc.h
@@ -109,7 +109,7 @@ struct drm_format_info {
 		 * information from their drm_mode_config.get_format_info hook
 		 * if they want the core to be validating the pitch.
 		 */
-		u8 char_per_block[DRM_FORMAT_MAX_PLANES];
+		u32 char_per_block[DRM_FORMAT_MAX_PLANES];
 	};
 
 	/**
